+++ sdk/fabric/arm-fabric/src/index.ts
@@ -16,10 +16,13 @@ export {
   Origin,
   KnownActionType,
   ActionType,
+  ErrorResponse,
+  ErrorDetail,
+  ErrorAdditionalInfo,
   FabricCapacity,
   FabricCapacityProperties,
-  ProvisioningState,
   KnownProvisioningState,
+  ProvisioningState,
   KnownResourceState,
   ResourceState,
   CapacityAdministration,
@@ -39,6 +42,8 @@ export {
   CheckNameAvailabilityReason,
   RpSkuDetailsForExistingResource,
   RpSkuDetailsForNewResource,
+  RpUsageAndQuotaDetailsForExistingResource,
+  KnownVersions,
 } from "./models/index.js";
 export {
   FabricClientOptionalParams,
@@ -53,7 +58,11 @@ export {
   FabricCapacitiesCheckNameAvailabilityOptionalParams,
   FabricCapacitiesListSkusForCapacityOptionalParams,
   FabricCapacitiesListSkusOptionalParams,
+  FabricCapacitiesListUsagesOptionalParams,
   OperationsListOptionalParams,
 } from "./api/index.js";
-export { FabricCapacitiesOperations, OperationsOperations } from "./classic/index.js";
+export {
+  FabricCapacitiesOperations,
+  OperationsOperations,
+} from "./classic/index.js";
 export { PageSettings, ContinuablePage, PagedAsyncIterableIterator };