+++ sdk/fabric/arm-fabric/CHANGELOG.md
@@ -1,15 +1,29 @@
 # Release History
-
-## 1.0.1 (Unreleased)
-
+    
+## 2.0.0 (2024-12-08)
+    
 ### Features Added
+
+  - Added operation FabricCapacitiesOperations.listUsages
+  - Added Interface ErrorAdditionalInfo
+  - Added Interface ErrorDetail
+  - Added Interface ErrorResponse
+  - Added Interface FabricCapacitiesListUsagesOptionalParams
+  - Added Interface RpUsageAndQuotaDetailsForExistingResource
+  - Added Enum KnownVersions
+  - Enum KnownOrigin has a new value "user,system"
+  - Enum KnownOrigin has a new value system
+  - Enum KnownOrigin has a new value user
+  - Enum KnownRpSkuTier has a new value fabric
 
 ### Breaking Changes
 
-### Bugs Fixed
-
-### Other Changes
-
+  - Enum KnownOrigin no longer has value System
+  - Enum KnownOrigin no longer has value User
+  - Enum KnownOrigin no longer has value UserSystem
+  - Enum KnownRpSkuTier no longer has value Fabric
+    
+    
 ## 1.0.0 (2024-10-14)
 
 ### Features Added